ecomsite/requirements.txt
asgiref==3.8.1
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
Django==5.1.1
djangorestframework==3.15.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.1.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.31.5
MarkupSafe==2.1.5
msgpack==1.1.0
packaging==24.1
pluggy==1.5.0
psutil==6.0.0
pycparser==2.22
pytest==8.3.3
pytest-django==4.9.0
pywin32==306
pyzmq==26.2.0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
